feat(cleanup): move CopyToClipboardButton to sensible spot
