+++ src/color/controls/CopyToClipboardButton.tsx
